# docs/openapi.yaml
...
openapi: "3.0.1"
info:
  title: "API Системы Управления Банковскими Картами"
  description: "REST API для управления банковскими картами с функциями:\n\n- Ауте\
    нтификация и авторизация пользователей\n- Управление картами (создание, просмо\
    тр, блокировка)\n- Переводы между картами\n- Администрирование системы\n\n**Ро\
    ли пользователей:**\n- USER: базовые операции с собственными картами\n- ADMIN:\
    \ полный доступ к системе\n\n**Безопасность:**\n- JWT токены для аутентификаци\
    и\n- Шифрование номеров карт\n- Маскирование чувствительных данных\n"
  contact:
    name: "Abu"
    url: "https://github.com/DesBasito"
    email: "out1of1mind1exception@gmail.com"
  version: "1.0.0"
servers:
- url: "http://localhost:8080/api/v1"
  description: "Локальный сервер разработки"
security:
- Bearer Authentication: []
...
paths:
...
  /card-applications:
    post:
      tags:
      - "Заявки на карты"
      summary: "Создать заявку на карту"
      description: "Создание заявки на новую карту пользователем"
      operationId: "createApplication"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardApplicationRequest"
        required: true
      responses:
        "200":
          description: "Заявка создана"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardApplicationDto"
      security:
      - Bearer Authentication: []
  /card-applications/{id}/reject:
    post:
      tags:
      - "Заявки на карты"
      summary: "Отклонить заявку (админ)"
      description: "Отклонение заявки администратором"
      operationId: "rejectApplication"
      parameters:
      - name: "id"
        in: "path"
        description: "ID заявки"
        required: true
        schema:
          type: "integer"
          format: "int64"
      - name: "reason"
        in: "query"
        description: "Причина отклонения"
        required: false
        schema:
          type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardApplicationDto"
      security:
      - Bearer Authentication: []
  /card-applications/{id}/cancel:
    post:
      tags:
      - "Заявки на карты"
      summary: "Отменить заявку"
      description: "Отмена заявки пользователем (только в статусе PENDING)"
      operationId: "cancelApplication"
      parameters:
      - name: "id"
        in: "path"
        description: "ID заявки"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardApplicationDto"
      security:
      - Bearer Authentication: []
  /card-applications/{id}/approve:
    post:
      tags:
      - "Заявки на карты"
      summary: "Одобрить заявку (админ)"
      description: "Одобрение заявки и создание карты"
      operationId: "approveApplication"
      parameters:
      - name: "id"
        in: "path"
        description: "ID заявки"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "Заявка одобрена, карта создана"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardDto"
        "400":
          description: "Заявка уже обработана"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CardDto"
      security:
      - Bearer Authentication: []
...
  /cards/my:
    get:
      tags:
      - "Карты"
      summary: "Получить все карты пользователя"
      description: "Получение списка карт текущего пользователя с пагинацией"
      operationId: "getMyCards"
      parameters:
      - name: "pageable"
        in: "query"
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: "Список карт получен"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/Page"
      security:
      - Bearer Authentication: []
...
  /card-applications/my:
    get:
      tags:
      - "Заявки на карты"
      summary: "Получить мои заявки"
      description: "Получение списка заявок текущего пользователя"
      operationId: "getMyApplications"
      parameters:
      - name: "pageable"
        in: "query"
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageCardApplicationDto"
      security:
      - Bearer Authentication: []
  /card-applications/admin/all:
    get:
      tags:
      - "Заявки на карты"
      summary: "Получить все заявки (админ)"
      description: "Получение всех заявок в системе (только для администраторов)"
      operationId: "getAllApplications"
      parameters:
      - name: "status"
        in: "query"
        description: "Статус заявки"
        required: false
        schema:
          type: "string"
      - name: "pageable"
        in: "query"
        required: true
        schema:
          $ref: "#/components/schemas/Pageable"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/PageCardApplicationDto"
      security:
      - Bearer Authentication: []
components:
  schemas:
...
    CardDto:
      required:
      - "balance"
      - "cardNumber"
      - "createdAt"
      - "expiryDate"
      - "ownerName"
      - "status"
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
        cardNumber:
          type: "string"
        ownerName:
          type: "string"
        ownerId:
          type: "integer"
          format: "int64"
        expiryDate:
          type: "string"
          format: "date"
        status:
          type: "string"
        type:
          type: "string"
        balance:
          type: "number"
        createdAt:
          type: "string"
          format: "date-time"
        updatedAt:
          type: "string"
          format: "date-time"
...
    CardApplicationRequest:
      required:
      - "cardType"
      type: "object"
      properties:
        cardType:
          type: "string"
          description: "Тип карты"
          example: "DEBIT"
          enum:
          - "DEBIT"
          - "CREDIT"
          - "VIRTUAL"
          - "PREPAID"
        comment:
          type: "string"
          description: "Комментарий к заявке"
          example: "Основная карта для зарплаты"
      description: "Запрос на создание карты"
    CardApplicationDto:
      type: "object"
      properties:
        id:
          type: "integer"
          description: "ID заявки"
          format: "int64"
        userId:
          type: "integer"
          description: "ID пользователя"
          format: "int64"
        userName:
          type: "string"
          description: "Имя пользователя"
        cardType:
          type: "string"
          description: "Тип карты"
        comment:
          type: "string"
          description: "Комментарий к заявке"
        status:
          type: "string"
          description: "Статус заявки"
          enum:
          - "PENDING"
          - "APPROVED"
          - "REJECTED"
          - "CANCELLED"
        createdAt:
          type: "string"
          description: "Дата создания заявки"
          format: "date-time"
        processedAt:
          type: "string"
          description: "Дата обработки заявки"
          format: "date-time"
      description: "Заявка на создание карты"
...
    Pageable:
      type: "object"
      properties:
        page:
          minimum: 0
          type: "integer"
          format: "int32"
        size:
          minimum: 1
          type: "integer"
          format: "int32"
        sort:
          type: "array"
          items:
            type: "string"
...
    PageableObject:
      type: "object"
      properties:
        offset:
          type: "integer"
          format: "int64"
        sort:
          $ref: "#/components/schemas/SortObject"
        paged:
          type: "boolean"
        unpaged:
          type: "boolean"
        pageNumber:
          type: "integer"
          format: "int32"
        pageSize:
          type: "integer"
          format: "int32"
    SortObject:
      type: "object"
      properties:
        empty:
          type: "boolean"
        unsorted:
          type: "boolean"
        sorted:
          type: "boolean"
...
    Page:
      type: "object"
      properties:
        totalPages:
          type: "integer"
          format: "int32"
        totalElements:
          type: "integer"
          format: "int64"
        first:
          type: "boolean"
        last:
          type: "boolean"
        size:
          type: "integer"
          format: "int32"
        content:
          type: "array"
          items:
            type: "object"
        number:
          type: "integer"
          format: "int32"
        sort:
          $ref: "#/components/schemas/SortObject"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        numberOfElements:
          type: "integer"
          format: "int32"
        empty:
          type: "boolean"
...
    PageCardApplicationDto:
      type: "object"
      properties:
        totalPages:
          type: "integer"
          format: "int32"
        totalElements:
          type: "integer"
          format: "int64"
        first:
          type: "boolean"
        last:
          type: "boolean"
        size:
          type: "integer"
          format: "int32"
        content:
          type: "array"
          items:
            $ref: "#/components/schemas/CardApplicationDto"
        number:
          type: "integer"
          format: "int32"
        sort:
          $ref: "#/components/schemas/SortObject"
        pageable:
          $ref: "#/components/schemas/PageableObject"
        numberOfElements:
          type: "integer"
          format: "int32"
        empty:
          type: "boolean"
  securitySchemes:
    Bearer Authentication:
      type: "http"
      description: "Enter JWT token in format: Bearer {token}"
      scheme: "bearer"
      bearerFormat: "JWT"